# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: E-Commerce Order Service API
  description: |
    e-커머스 상품 주문 서비스 API 명세서
    
    ## 주요 기능
    - 상품 조회
    - 주문 및 결제
    - 포인트 충전 및 조회
    
    ## 제약사항
    - 재고 관리: 동시성 이슈 고려
    - 포인트 차감: 트랜잭션 처리 필수
    - 데이터 플랫폼 전송: 주문 성공 시 외부 시스템으로 전송
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: Products
    description: 상품 관련 API
  - name: Orders
    description: 주문 및 결제 관련 API
  - name: Points
    description: 포인트 관련 API

paths:
  /api/products:
    get:
      tags:
        - Products
      summary: 상품 목록 조회
      description: |
        전체 상품 목록을 조회합니다.
        조회 시점의 정확한 재고 수량을 반환합니다.
      operationId: getProducts
      parameters:
        - name: page
          in: query
          description: 페이지 번호 (0부터 시작)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: 페이지 크기
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 상품 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/products/{productId}:
    get:
      tags:
        - Products
      summary: 상품 상세 조회
      description: 특정 상품의 상세 정보를 조회합니다.
      operationId: getProduct
      parameters:
        - name: productId
          in: path
          required: true
          description: 상품 ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 상품 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: 상품을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/products/search:
    get:
      tags:
        - Products
      summary: 상품 검색
      description: |
        상품명을 기준으로 검색합니다.
        검색 이벤트는 Kafka를 통해 데이터 플랫폼으로 전달됩니다.
      operationId: searchProducts
      parameters:
        - name: keyword
          in: query
          required: true
          description: 검색어
          schema:
            type: string
        - name: X-User-Id
          in: header
          required: false
          description: 검색 사용자 ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 검색 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          description: 잘못된 검색어
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/points/charge:
    post:
      tags:
        - Points
      summary: 포인트 충전
      description: |
        사용자의 포인트를 충전합니다.
        충전 금액은 양수여야 합니다.
      operationId: chargePoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PointChargeRequest'
      responses:
        '200':
          description: 포인트 충전 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointResponse'
        '400':
          description: 잘못된 요청 (금액이 0 이하)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 사용자를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/points/{userId}:
    get:
      tags:
        - Points
      summary: 포인트 조회
      description: 특정 사용자의 현재 포인트 잔액을 조회합니다.
      operationId: getPoint
      parameters:
        - name: userId
          in: path
          required: true
          description: 사용자 ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 포인트 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointResponse'
        '404':
          description: 사용자를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/orders:
    post:
      tags:
        - Orders
      summary: 주문 및 결제
      description: |
        상품을 주문하고 포인트로 결제를 수행합니다.
        
        ## 처리 과정
        1. 재고 확인 (동시성 고려)
        2. 총 주문 금액 계산
        3. 포인트 잔액 확인
        4. 재고 차감 및 포인트 차감 (트랜잭션)
        5. 주문 정보 생성
        6. 데이터 플랫폼으로 주문 정보 전송 (비동기)
        
        ## 제약사항
        - 재고가 부족하면 주문 실패
        - 포인트가 부족하면 주문 실패
        - 동시 주문 시 재고 관리 정확성 보장
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: 주문 및 결제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: 잘못된 요청 (수량이 0 이하, 빈 주문 등)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 사용자 또는 상품을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 재고 부족 또는 포인트 부족
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - price
        - stockQuantity
      properties:
        id:
          type: integer
          format: int64
          description: 상품 ID
          example: 1
        name:
          type: string
          description: 상품명
          example: "노트북"
        price:
          type: integer
          format: int64
          description: 상품 가격 (원)
          example: 1000000
        stockQuantity:
          type: integer
          description: 잔여 수량
          example: 10

    ProductResponse:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/Product'

    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        page:
          type: integer
          description: 현재 페이지 (0부터 시작)
          example: 0
        size:
          type: integer
          description: 페이지 크기
          example: 20
        totalElements:
          type: integer
          description: 전체 상품 수
          example: 100
        totalPages:
          type: integer
          description: 전체 페이지 수
          example: 5

    PointChargeRequest:
      type: object
      required:
        - userId
        - amount
      properties:
        userId:
          type: integer
          format: int64
          description: 사용자 ID
          example: 1
        amount:
          type: integer
          format: int64
          description: 충전할 금액 (원, 양수)
          minimum: 1
          example: 50000

    PointResponse:
      type: object
      required:
        - userId
        - balance
      properties:
        userId:
          type: integer
          format: int64
          description: 사용자 ID
          example: 1
        balance:
          type: integer
          format: int64
          description: 현재 잔액 (원)
          example: 150000

    OrderItemRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: integer
          format: int64
          description: 상품 ID
          example: 1
        quantity:
          type: integer
          description: 주문 수량
          minimum: 1
          example: 2

    OrderRequest:
      type: object
      required:
        - userId
        - items
      properties:
        userId:
          type: integer
          format: int64
          description: 사용자 ID
          example: 1
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'
          description: 주문할 상품 목록

    OrderItem:
      type: object
      required:
        - productId
        - productName
        - quantity
        - price
      properties:
        productId:
          type: integer
          format: int64
          description: 상품 ID
          example: 1
        productName:
          type: string
          description: 상품명
          example: "노트북"
        quantity:
          type: integer
          description: 주문 수량
          example: 2
        price:
          type: integer
          format: int64
          description: 단가 (원)
          example: 1000000

    OrderResponse:
      type: object
      required:
        - orderId
        - userId
        - totalAmount
        - remainingBalance
        - status
        - items
        - createdAt
      properties:
        orderId:
          type: integer
          format: int64
          description: 주문 ID
          example: 1
        userId:
          type: integer
          format: int64
          description: 사용자 ID
          example: 1
        totalAmount:
          type: integer
          format: int64
          description: 총 주문 금액 (원)
          example: 2000000
        remainingBalance:
          type: integer
          format: int64
          description: 결제 후 잔액 (원)
          example: 500000
        status:
          type: string
          enum:
            - COMPLETED
            - FAILED
          description: 주문 상태
          example: "COMPLETED"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          description: 주문 상품 목록
        createdAt:
          type: string
          format: date-time
          description: 주문 생성 시간
          example: "2024-01-01T00:00:00Z"

    ErrorResponse:
      type: object
      required:
        - errorCode
        - message
        - timestamp
      properties:
        errorCode:
          type: string
          description: 에러 코드
          example: "PRODUCT_NOT_FOUND"
        message:
          type: string
          description: 에러 메시지
          example: "상품을 찾을 수 없습니다."
        timestamp:
          type: string
          format: date-time
          description: 에러 발생 시간
          example: "2024-01-01T00:00:00Z"